Move class into file
