+++ lib/fluent/engine.rb
@@ -22,6 +22,7 @@ require 'fluent/root_agent'
 require 'fluent/time'
 require 'fluent/system_config'
 require 'fluent/plugin'
+require 'fluent/fluent_log_event_router'
 
 module Fluent
   class EngineClass
@@ -31,7 +32,7 @@ module Fluent
       @engine_stopped = false
       @_worker_id = nil
 
-      @log_event_router = nil
+      @fluent_log_event_router = nil
       @log_event_verbose = false
 
       @suppress_config_dump = false
@@ -110,9 +111,9 @@ module Fluent
     def configure(conf)
       @root_agent.configure(conf)
 
-      @log_event_router = FluentLogEventEmitter.build(@root_agent)
+      @fleunt_log_event_router = FluentLogEventRouter.build(@root_agent)
 
-      if @log_event_router
+      if @fleunt_log_event_router
         $log.enable_event(true)
       end
 
@@ -151,8 +152,8 @@ module Fluent
         $log.info "starting fluentd worker", pid: Process.pid, ppid: Process.ppid, worker: worker_id
         start
 
-        if @log_event_router
-          @log_event_router.start
+        if @fleunt_log_event_router
+          @fleunt_log_event_router.start
         end
 
         $log.info "fluentd worker is now running", worker: worker_id
@@ -168,15 +169,15 @@ module Fluent
       unless @log_event_verbose
         $log.enable_event(false)
 
-        if @log_event_router
-          @log_event_router.graceful_stop
+        if @fleunt_log_event_router
+          @fleunt_log_event_router.graceful_stop
         end
       end
       $log.info "shutting down fluentd worker", worker: worker_id
       shutdown
 
-      if @log_event_router
-        @log_event_router.stop
+      if @fleunt_log_event_router
+        @fleunt_log_event_router.stop
       end
     end
 
@@ -186,8 +187,8 @@ module Fluent
     end
 
     def push_log_event(tag, time, record)
-      if @log_event_router
-        @log_event_router.emit_event([tag, time, record])
+      if @fleunt_log_event_router
+        @fleunt_log_event_router.emit_event([tag, time, record])
       end
     end
 
@@ -206,103 +207,6 @@ module Fluent
 
     def shutdown
       @root_agent.shutdown
-    end
-  end
-
-  class FluentLogEventEmitter
-    LOG_EMIT_INTERVAL = 0.1
-
-    def self.build(root_agent)
-      log_event_router = nil
-
-      begin
-        log_event_agent = root_agent.find_label(Fluent::Log::LOG_EVENT_LABEL)
-        log_event_router = log_event_agent.event_router
-
-        # suppress mismatched tags only for <label @FLUENT_LOG> label.
-        # it's not suppressed in default event router for non-log-event events
-        log_event_router.suppress_missing_match!
-
-        log_event_router = log_event_router
-
-        unmatched_tags = Fluent::Log.event_tags.select{|t| !log_event_router.match?(t) }
-        unless unmatched_tags.empty?
-          $log.warn "match for some tags of log events are not defined (to be ignored)", tags: unmatched_tags
-        end
-
-      rescue ArgumentError # ArgumentError "#{label_name} label not found"
-        # use default event router if <label @FLUENT_LOG> is missing in configuration
-        log_event_router = root_agent.event_router
-
-        if Fluent::Log.event_tags.any?{|t| log_event_router.match?(t) }
-          log_event_router = log_event_router
-
-          unmatched_tags = Fluent::Log.event_tags.select{|t| !log_event_router.match?(t) }
-          unless unmatched_tags.empty?
-            $log.warn "match for some tags of log events are not defined (to be ignored)", tags: unmatched_tags
-          end
-        end
-      end
-
-      if log_event_router
-        FluentLogEventEmitter.new(log_event_router)
-      else
-        nil # no need to create fluent log event emitter
-      end
-    end
-
-    def initialize(event_router)
-      @event_router = event_router
-      @thread = nil
-      @graceful_stop = false
-      @stopped = false
-      @event_queue = []
-    end
-
-    def start
-      @thread = Thread.new do
-        $log.disable_events(Thread.current)
-
-        loop do
-          sleep(LOG_EMIT_INTERVAL)
-
-          break if @stop
-          break if @graceful_stop && @event_queue.empty?
-
-          next if @event_queue.empty?
-
-          # NOTE: thead-safe of slice! depends on GVL
-          events = @event_queue.slice!(0..-1)
-          next if events.empty?
-
-          events.each do |tag, time, record|
-            begin
-              @event_router.emit(tag, time, record)
-            rescue => e
-              # This $log.error doesn't emit log events, because of `$log.disable_events(Thread.current)` above
-              $log.error "failed to emit fluentd's log event", tag: tag, event: record, error: e
-            end
-          end
-        end
-      end
-
-      @thread.abort_on_exception = true
-    end
-
-    def stop
-      @stop = true
-      # there is no problem calling Thread#join multiple times.
-      @thread.join
-    end
-
-    def graceful_stop
-      # to make sure to emit all log events into router, before shutting down
-      @graceful_stop = true
-      @thread.join
-    end
-
-    def emit_event(event)
-      @event_queue.push(event)
     end
   end
